requirements.txt
# Pure Python dependencies for photo/movie organization script

# Image processing and EXIF extraction
Pillow>=10.0.0

# Video/audio metadata extraction (pure Python)
mutagen>=1.47.0

# Testing framework
pytest>=7.0.0

